Moved insert_uniques and update_many method from advanced to core module
